feat: Add user_id validation in createDonation
The code changes in `donationsController.js` add user_id validation in the `createDonation` function. It checks if the user exists before inserting a new donation into the database. If the user is not found, an error message is returned. This update improves the reliability of the donation creation process.

Note: This suggested commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes made.