feat: Add image field to createProduct and updateProduct functions
This commit modifies the `createProduct` and `updateProduct` functions in the `productsController.js` file to include the `image` field in the database queries. The `image` field is now being inserted and updated along with other product details. This update allows the application to store and retrieve images associated with each product, enhancing the visual representation of the products.